Poison and tracker fix

+++ src/com/crackedcarrot/Creature.java
@@ -49,7 +49,7 @@ public class Creature extends Sprite{
     public float creatureFrozenTime;
     public float creatureFrozenAmount = 1;
     public float creaturePoisonTime;
-    public int creaturePoisonDamage;
+    public float creaturePoisonDamage;
 	// This is the base rgb colors for this creature
 	protected float rDefault = 1;
 	protected float gDefault = 1;
@@ -362,9 +362,9 @@ public class Creature extends Sprite{
 	public void affectWithPoison(int poisonTime, float poisonDamage) {
 		if (!this.creaturePoisonResistant) {
 			if ( this.creaturePoisonDamage > 0 && this.creaturePoisonTime > 0 )
-				this.creaturePoisonDamage = (int)(poisonDamage + (this.creaturePoisonDamage * this.creaturePoisonTime) / poisonTime);
+				this.creaturePoisonDamage = ((poisonDamage + (this.creaturePoisonDamage * this.creaturePoisonTime)) / poisonTime);
 			else
-				this.creaturePoisonDamage = (int)(poisonDamage);
+				this.creaturePoisonDamage = poisonDamage;
 			this.creaturePoisonTime = poisonTime;
 		}
 	}
